Moved Fetch literal functionality to separate file
